requirements.txt
uvicorn
fastapi
firebase-admin
pymongo
requests
wrangler
requests
python-dotenv